requirements.txt
Django==3.1.3
djangorestframework==3.12.2
pylint==2.6.0
python-dotenv==0.15.0
drf-yasg==1.20.0
swagger-spec-validator==2.7.3